UI list units: UIListSlot::Draw 100%, HamListRibbon Start/EndFrame 100%, accessor outline fix
Key improvements:
- UIListSlot::Draw 80.8% → 100%: moved DisabledAlphaScale() and ParentList()
  from inline header to UIListWidget.cpp (target didn't inline these accessors)
- HamListRibbon::StartFrame/EndFrame 93% → 100%: removed redundant OR guards
  where inner conditions were collectively exhaustive
- UIListLabel::Load 86% → 99.2%: backward-compat String read + BinStreamRev
- UIListLabelElement::Draw 54% → 73.3%: XZ plane vectors, unsigned loops
- HamNavList::DrawShowing 64.5% → 90.7%: structural fixes
- UIListDir::DrawWidgets: foreach→do-while, scope narrowing, bool caching
- Permuter: m2c guidance, new patterns, beam search improvements
- Renamed UIListElementDrawState unk fields to mScaleX/mScaleY/mScaleZ
- Added accessor_outline and redundant_guard_elimination to synthesis roadmap

+++ src/lazer/game/Game.cpp
@@ -673,7 +673,7 @@ void Game::LoadNewSong(Symbol s1, Symbol s2) {
 
 void Game::PauseForSkeletonLoss() {
     if (!mPaused) {
-        int gestureVal = TheGestureMgr->GetPauseOnSkeletonLossMode();
+        short gestureVal = TheGestureMgr->GetPauseOnSkeletonLossMode();
         if (gestureVal != 0 && gestureVal != 1 && !TheSynth->HasPendingVoices()
             && !TheUI->InTransition()) {
             static Message pauseOnSkeletonLossMsg("pause_on_skeleton_loss");
